+++ osl/source_recon/__init__.py
@@ -1,3 +1,3 @@
 from .batch import *  # noqa: F401, F403
-from .rhino.fsl_wrappers import setup_fsl  # noqa: F401, F403
+from .rhino.fsl_utils import setup_fsl  # noqa: F401, F403
 from .wrappers import find_template_subject  # noqa: F401, F403
