+++ osl/source_recon/batch.py
@@ -127,7 +127,7 @@ def run_src_chain(
     flag : bool
         Flag indicating whether source reconstruction was successful.
     """
-    rhino.fsl_wrappers.check_fsl()
+    rhino.fsl_utils.check_fsl()
 
     # Directories
     src_dir = validate_outdir(src_dir)
@@ -251,7 +251,7 @@ def run_src_batch(
     flags : list of bool
         Flags indicating whether coregistration was successful.
     """
-    rhino.fsl_wrappers.check_fsl()
+    rhino.fsl_utils.check_fsl()
 
     # Directories
     src_dir = validate_outdir(src_dir)
