+++ osl/source_recon/rhino/__init__.py
@@ -6,5 +6,5 @@ from .coreg import *  # noqa: F401, F403
 from .polhemus import *  # noqa: F401, F403
 from .forward_model import *  # noqa: F401, F403
 from .surfaces import *  # noqa: F401, F403
-from .fsl_wrappers import *  # noqa: F401, F403
+from .fsl_utils import *  # noqa: F401, F403
 from . import utils  # noqa: F401, F403
